Strip header value and treat empty values as undefined headers.
RFC2616 states: The field-content does not include any leading or
trailing LWS: linear white space occurring before the first
non-whitespace character of the field-value or after the last
non-whitespace character of the field-value. This change brings
mochiweb into compliance with that statement by stripping whitespace
from the header values. Additionally headers that are specified with
no non-whitespace characters are treated as undefined headers.