Addmissing Opts to handle_invalid_msg_request
Also needed toc hange the unit test.  the unit test was testing a bad header ... but then expecting a 200 OK response ... but it is invalid?

The test no longer exists in this form - so I'm not sure whether it is valid to change this.  I can't find direct evidence the test was previously wrong.

Perhaps I have misunderstoof the purpose of the test.  https://github.com/basho/mochiweb/commit/5a3d51172814f309b132f818cb5902419245c5ef.  Does this imply that the test is expecting to handle another message correctly even if interrupted by a partial header?