Fix warning, add error log on Other
It appears the test was originaly menat to work as the "Other" case would be hit, and then when hitting the Other case - the code would previously just loop round without adding a header.

This is no longer the case (i.e. even if the original 'Other' handling is provided).