Add support for 431 response code
Headers must be smaller than recbuf, but in some cases they won't be.  Simply returning 400 Bad Request with no indication is not helpful - header code 431 is designed for this - https://tools.ietf.org/html/rfc6585#section-5

This response code is not supported in the erlang httpd_util, so need to override the mapping of codes to response messages for this code to provide the supported return.